Change stracture for testing environment and update test with helper function
